backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
sqlmodel>=0.0.8
alembic>=1.12.0
python-dotenv>=1.0.0
passlib[argon2]>=1.7.4
python-jose[cryptography]>=3.3.0
httpx>=0.24.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
requests>=2.31.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.5
slowapi>=0.1.9
aiosqlite>=0.19.0
motor>=3.3.0
pymongo>=4.5.0
aiohttp>=3.9.0
pytz>=2023.3
